+++ src/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import JudgeInfo from './components/JudgeInfo/JudgeInfo';
+
+import './App.css';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <JudgeInfo />
+      </div>
+    );
+  }
+}
+
+export default App;
