+++ src/App.test.js
@@ -1,10 +1,10 @@
 /* eslint-disable */
 import React from 'react';
 import ReactDOM from 'react-dom';
-import JudgeInput from './JudgeInput';
+import App from './App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<JudgeInput />, div);
+  ReactDOM.render(<App />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
