--- src/JudgeInput.css
@@ -1,3 +1,0 @@
-.JudgeName {
-  text-align: center;
-}
